Move stories in each packages
