+++ packages/react-components/react-grid-layout/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
