+++ packages/react-components/react-responsive-image/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
