+++ packages/react-components/react-responsive-image/package.json
@@ -26,7 +26,8 @@
   },
   "devDependencies": {
     "@types/react": "^16.9.48",
-    "@types/react-dom": "^16.9.8"
+    "@types/react-dom": "^16.9.8",
+    "@wbe/fake-data-utils": "^1.3.1-alpha.1"
   },
   "keywords": [
     "react",
