+++ packages/react-components/react-video-player/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
