+++ packages/react-components/react-video-player/package.json
@@ -26,6 +26,7 @@
   "devDependencies": {
     "@types/react": "^16.9.48",
     "@types/react-dom": "^16.9.8",
+    "@wbe/fake-data-utils": "^1.3.1-alpha.1",
     "typescript": "^4.0.2"
   },
   "keywords": [
