+++ packages/react-components/react-video-player/src/index.tsx
@@ -15,7 +15,7 @@ export enum EVideoType {
 /**
  * VideoPlayer Props
  */
-interface IProps {
+export interface IProps {
   /**
    * Choose video type <native | vimeo | youtube>
    */
