+++ packages/react-hooks/use-async-effect/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
