+++ packages/react-hooks/use-bounding-client-rect/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
