+++ packages/react-hooks/use-bounding-client-rect/stories/useBoundingClientRect.stories.tsx
@@ -1,20 +1,15 @@
+// @ts-ignore
+import useBoundingClientRect, { EListener } from "../src";
 import React, { useRef } from "react";
-import useBoundingClientRect, {
-  EListener
-} from "@wbe/use-bounding-client-rect";
-import README from "@wbe/use-bounding-client-rect/README.md";
-import "../../global-style.css";
+import "../../../../storybook/global-style.css";
 
 const storyName = "use-bounding-client-rect";
 
 export const App = () => {
-  const elementRef = useRef(null);
-  const rect = useBoundingClientRect(
-    elementRef,
-    EListener.ON_SCROLL_AND_RESIZE
-  );
+  const rootRef = useRef(null);
+  const rect = useBoundingClientRect(rootRef, EListener.ON_SCROLL_AND_RESIZE);
   return (
-    <div ref={elementRef}>
+    <div ref={rootRef}>
       <p> Resize your browser and check element properties change.</p>
       <pre>{JSON.stringify(rect, null, 2)}</pre>
     </div>
