+++ packages/react-hooks/use-document-scroll-top/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
