+++ packages/react-hooks/use-document-scroll-top/tsconfig.json
@@ -4,6 +4,5 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "exclude": ["node_modules", "dist", "test"],
-  "include": ["src/*"]
+  "exclude": ["node_modules", "dist", "test", "stories"]
 }
