+++ packages/react-hooks/use-is-in-viewport/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
