+++ packages/react-hooks/use-is-in-viewport/tsconfig.json
@@ -8,5 +8,5 @@
     { "path": "../../react-hooks/use-bounding-client-rect" },
     { "path": "../../react-hooks/use-window-size" }
   ],
-  "exclude": ["node_modules", "dist", "test"]
+  "exclude": ["node_modules", "dist", "test", "stories"]
 }
