+++ packages/react-hooks/use-responsive-image-data/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+stories
+test
