+++ packages/react-hooks/use-responsive-image-data/package.json
@@ -21,11 +21,12 @@
     "react": "^16.13.1"
   },
   "devDependencies": {
+    "@types/react": "^16.9.48",
     "@testing-library/react": "^11.0.2",
     "@testing-library/react-hooks": "^3.4.1",
-    "@types/react": "^16.9.48",
+    "react-test-renderer": "^16.13.1",
     "@wbe/fake-data-utils": "^1.3.1-alpha.1",
-    "react-test-renderer": "^16.13.1"
+    "@wbe/use-window-size": "^1.3.1-alpha.1"
   },
   "keywords": [
     "react",
