+++ packages/react-hooks/use-window-size/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+stories
+test
