+++ packages/utils/fake-data-utils/.npmignore
@@ -2,3 +2,5 @@
 .cache
 node_modules
 src
+test
+stories
