+++ packages/utils/fake-data-utils/package.json
@@ -24,5 +24,9 @@
   "publishConfig": {
     "access": "public"
   },
-  "gitHead": "6685e8d20f475c4fbd0ef670700a6f71add3a09e"
+  "gitHead": "6685e8d20f475c4fbd0ef670700a6f71add3a09e",
+  "devDependencies": {
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
+  }
 }
