+++ storybook/.storybook/customTheme.js
@@ -3,5 +3,13 @@ import { create } from "@storybook/theming/create";
 export default create({
   base: "light",
   brandTitle: "📦 Libraries",
-  brandUrl: " https://github.com/willybrauner/libraries"
+  brandUrl: " https://github.com/willybrauner/libraries",
+
+  // Typography
+  fontBase: '"Open Sans", sans-serif',
+  fontCode: "monospace",
+
+  // Text colors
+  textColor: "black",
+  textInverseColor: "rgba(255,255,255,0.9)"
 });
