+++ storybook/package.json
@@ -10,18 +10,6 @@
   },
   "author": "Willy Brauner",
   "license": "ISC",
-  "dependencies": {
-    "@wbe/fake-data-utils": "^1.3.1-alpha.1",
-    "@wbe/react-grid-layout": "^1.3.1-alpha.1",
-    "@wbe/react-responsive-image": "^1.3.1-alpha.1",
-    "@wbe/react-video-player": "^1.3.1-alpha.1",
-    "@wbe/use-async-effect": "^1.3.0",
-    "@wbe/use-bounding-client-rect": "^1.3.0",
-    "@wbe/use-document-scroll-top": "^1.3.0",
-    "@wbe/use-is-in-viewport": "^1.3.1-alpha.1",
-    "@wbe/use-responsive-image-data": "^1.3.1-alpha.1",
-    "@wbe/use-window-size": "^1.3.1-alpha.1"
-  },
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@storybook/addon-essentials": "^6.0.21",
