+++ tsconfig.json
@@ -19,5 +19,5 @@
     // pass to true for disable this check.
     "skipLibCheck": false
   },
-  "exclude": ["node_modules", "dist", "test"]
+  "exclude": ["node_modules", "dist", "test", "stories"]
 }
